Detect internal HTTPS redirects when detecting external sites
If a server address is entered with HTTP or no protocol, handle URLs with both HTTP and HTTPS as internal sites and do not open a browser.

It may be better to find some method of "following" redirects so we can determine exactly where the "real" address is located.That way we can handle redirects more complicated than HTTP -> HTTPS. If that seems too overboard, this should be enough to catch the situations where people don't type in a protocol.